requirements.txt
fastapi
pydantic
python-dotenv
httpx
asyncpg
qdrant-client
langchain-qdrant
langchain-core
langchain-google-genai
uvicorn
openai
tenacity